potential bug fixes

+++ reversi/led-matrix/Color.h
@@ -4,8 +4,8 @@ struct Color
 };
 
 //Percentage is percentage from c1 to c2
-Color interpolate(Color c1, Color c2, int percentage) {
-	double fraction = (double) percentage/100.0;
+Color interpolate(Color c1, Color c2, double percentage) {
+	double fraction = percentage/100.0;
 
 	Color c3;
 	c3.R = 0;
@@ -13,9 +13,9 @@ Color interpolate(Color c1, Color c2, int percentage) {
 	c3.B = 0;
 
 
-	c3.R = c1.R + (c2.R - c1.R)*fraction;
-	c3.G = c1.G + (c2.G - c1.G)*fraction;
-	c3.B = c1.B + (c2.B - c1.B)*fraction;
+	c3.R = c1.R + (double)(c2.R - c1.R)*fraction;
+	c3.G = c1.G + (double)(c2.G - c1.G)*fraction;
+	c3.B = c1.B + (double)(c2.B - c1.B)*fraction;
 
 	return c3;
 }
